Rollback from separating src and header files.
Vector.cpp gets compiled before the main file. Hence, the class method
written implemented in the src file will not be specialized for the
object (Vector<int>) that will be created by the main. There are ways to
get around this and tell the Vector.cpp to generate the function for int
after the main.cpp is compiled but they are not clean. After checking
clang's implementation I suppose writing everything in the header is ok.